Added display on demand row loading (virtual table)
On demand row loading removes any QTableView startup costs.

On demand row loading is done using the recommended canFetchMore/fetchMore pair
and returning the current loading count from rowCount (this causes the vertical
bar to shrink as new rows are loaded, but doesn't seem to be a way of avoding
that).

Note this is only on-demand from the point of view of displaying the rows, the
directory data is still loaded wholesome at application startup. But with this
implementation, in the future the directory data could be stored in a file or
database and loaded on demand too.

In addition, this has other benefits:
- Filtering is faster since only loaded rows are filtered. The worst case is
  when filtering for an inexistent word or scrolling the table to the last
  occurrence of the filtered word, where the whole source data needs to be
  traversed.
- Sorting is done by sorting the source data and invalidating the model, which
  is fast, rather than using proxy models.

Since now filtering is reasonably fast even on Raspberry Pi, the filter is now
updated on every key stroke instead of having to press enter.

Files:

README.md
- Updated with the new features

filefinder.py
- Implemented on demand row loading.
- Removed SortFilterProxyModel, since it doesn't work with on demand loading (
  beginInsertRow calls filterAcceptsRow which fails because there's no data
  yet). Not a big deal since implementing filtering is easy.
- Implemented sorting by sorting the original data and invalidating the model
- Implemented filtering by invalidating the model and skipping filtered out
  rows at on demand load time
- Rationalized path normalization, turns out Qt uses forward slashes, so only
  need to normpath when reading from Qt apis.

Tests:
- Ran on raspbian, verified user directory works, directories are traversed,
  verified rows are displayed on demand, filtered with the combobox values,
  double and right click work as expected.
- Ran on Windows, verified paths are normalized, verified directories are
  traversed, verified rows are displayed on demand, filtered with the combobox
  values, double and right click work as expected.