Switched from csv files to sqlite3 database in WAL mode, used Python logging
csv files are slow for sorting and for filtering with 1M entries, this commit
moves from csv files to a sqlite3 database, which is able to handle 1M files

On Raspberry Pi 2 with 1M files
- Changing the sorting order is immediate
- Adding a char to the filter is immediate as long as it doesn't need to scan
  the whole database to fill the table, in that case it can take a minute (the
  DB is 800MB and on a USB hard drive, 1M Files, 60K dirs).
On laptop both are almost immediate even if the filter string is not found.

The approach used is to enable sqlite3 WAL mode on the database (otherwise a
reader cannot be concurrent with a writer) with three connections:
- A reader connection for the TableView. The TableView has to be updated
  manually by updating the filter lineedit
- A reader connection to walk the db looking for diffs wrt the filesystem
  This cursor is refreshed explicitly as needed when the writer connection
  commits (the writer updates may be about new directories that need to be
  traversed)
- A writer connection to update the db with the changes

The current database schema is a simple table that replicates the paths in all
the entries, this makes the substring query simple without needing cross-table
joins, but it wastes space.

Early tests splitting paths into its own table reduce the database file to a 38%
of the current one (from 683MB to 261MB) and the empty-results query time to 50%.
Non-empty result query time is 10% slower.

The current UI is not very responsive when ingesting remote directories with
lots of files, ideally the db udpate should be moved to a different process,
which will also make progress towards a client/server approach.

Files:
filefinder.py:
- Added logging functions
- Adapted tablemodel to use a sqlite database, no longer does manual filtering
  or sorting in Python, rather it builds the necessary sort and filter sql
  clauses and redoes the sql query.
- Changed filtered row count to loaded row count in all places since there's no
  manual filtering anymore

README.md:

Tests:
- Verified db updates properly when files / dirs created
- Verified db updates properly when files / dirs are deleted
- Verified db updates properly when files / dirs are renamed/moved
- Verified queries can be done at the same time the db is being updated
- Verified the WAL file is eventually deleted
- Verified smart updates work
- Verified sorting, filtering works and it's fast on laptop with 1M files
  (filtering can be slow on Rasperry Pi when a full DB scan is needed)
- Verified creating a new db from scratch works