Fix for undhandled exception regression when a directory is deleted/renamed
The previous commit raised an unhandled exception instead of creating a dummy
directory entry when a traversed directory wasn't found in the database (note
that new directories should always have an entry in the database before
traversed, since the entry is created when traversing the parent).

Creating a dummy directory entry was wong (it was only there because other parts
of the code didn't handle unreacheable host properly) and caused database
entries to be deleted if the host wasn't reacheable (eg due to temporary network
errors).

Turns out that exception is also hit when a directory is removed from the
filesystem when traversing the parent, and the directory traversed when some
children are found.

Now the code does abort in the presence of temporary filesystem errors, so the
right thing to do in this case is to fall through and let the code delete the
children of the directory being traversed.

A better option could be to delete with a prefix query instead of child by
child, but that kind of query probably needs a full table scan which is
probably slower.

Files:
filefinder.py:
- Fall through when the directory being traversed is not in the database instead
  of rising an exception. The fall through code will then delete the children
  from the database

README.md:
- Added todo

Tests:
- Verified directories deleted/renamed in the filesystem are properly deleted
  from the database