Added Copy/Cut files to clipboard, database filepath command line argument
This commit adds support for copying and cutting files to the clipboard on
Windows and Linux (note, as is usual behavior, files are not deleted until the
paste is done in whatever target desktop program performs the paste since the
deletion depends on the target, not on the source).

Copy and cut are the usual ctrl+c and ctrl+v, which means that copying filepaths
is now changed to ctrl+shift+c

In addition, the program now takes the full or relative path and filename for
the database so multiple instances can be run indexing different directories on
different databases.

Files:

filefinder.py
- Added file copy and cut to clipboard support
- Added database command line argument.

README.md
- Added new copy/cut feature, changed running examples

Tests:
- Tested under Windows, Raspberry Pi with LXDE, verified both copy and cut work.
- Tested multiple insteances using multiple databases.